+++ lib/send_grid_mailer/engine.rb
@@ -9,6 +9,8 @@ module SendGridMailer
 
     initializer "initialize" do
       require_relative "./errors"
+      require_relative "./interceptors_handler"
+      require_relative "./interceptor/recipient_interceptor"
       require_relative "./logger"
       require_relative "./definition"
       require_relative "./mailer_base_ext"